requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0

# Experiment Tracking
wandb>=0.15.0

# Data Processing
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Streamlit for Labeling Interface
streamlit>=1.28.0
streamlit-image-select>=0.6.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Progress bars and utilities
tqdm>=4.65.0

# Development
pytest>=7.4.0
black>=23.0.0
